@@ -1,16 +1,16 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.0" />
-		<PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.15.1" />
-		<PackageReference Include="Otp.NET" Version="1.2.2" />
-		<PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.8.0" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="7.0.0" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="7.0.4" />
+		<PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.32.3" />
+		<PackageReference Include="Otp.NET" Version="1.3.0" />
+		<PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.32.3" />
 	</ItemGroup>
 	<ItemGroup>
 	  <ProjectReference Include="..\CoreFramework.Persistence\CoreFramework.Persistence.csproj" />
