@@ -1,4 +1,5 @@
-﻿using CoreFramework.Security.Entities;
+﻿using Application.Features.Auths.Dtos;
+using CoreFramework.Security.Entities;
 using CoreFramework.Security.JWT;
 using System;
 using System.Collections.Generic;
@@ -8,10 +9,11 @@ using System.Threading.Tasks;
 
 namespace Application.Services.AuthService
 {
-    public interface IAuthService
+    public interface IAuthService //: IInternalAuthentication
     {
-        public Task<AccessToken> CreateAccessToken(User user);
-        public Task<RefreshToken> CreateRefreshToken(User user, string ipAddress);
-        public Task<RefreshToken> AddRefreshToken(RefreshToken refreshToken);
+        Task PasswordResetAsnyc(string email);
+        Task<bool> VerifyResetTokenAsync(string resetToken, string userId);
+        Task<Token> LoginAsync(string usernameOrEmail, string password, int accessTokenLifeTime);
+        Task<Token> RefreshTokenLoginAsync(string refreshToken);
     }
 }
