@@ -1,5 +1,4 @@
-﻿using CoreFramework.Persistence.Repositories;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -7,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Domain.Entities
 {
-    internal class Test:Entity
+    internal class Test //:Entity
     {
         public int Id { get; set; }     
     }
