@@ -1,3 +1,3 @@
 using Application.Services.Repositories;using Persistence.Contexts;using Persistence.Repositories;using Microsoft.EntityFrameworkCore;using Microsoft.Extensions.Configuration;using Microsoft.Extensions.DependencyInjection;using Domain.Entities;
 
-namespace Persistence{    public static class PersistenceServiceRegistration    {        public static IServiceCollection AddPersistenceServices(this IServiceCollection services, IConfiguration configuration)        {            services.AddDbContext<SmsDbContext>(options => options.UseNpgsql("User ID=makin19;Password=Muhammed1453;Host=81.0.220.136;Port=5432;Database=SMS;"));            services.AddIdentityCore<User>().AddRoles<Role>().AddEntityFrameworkStores<SmsDbContext>();            services.AddScoped<IRefreshTokenRepository, RefreshTokenRepository>();            services.AddScoped<IOperationClaimRepository, OperationClaimRepository>();            services.AddScoped<IUserOperationClaimRepository, UserOperationClaimRepository>();            return services;        }    }}
+namespace Persistence{    public static class PersistenceServiceRegistration    {        public static IServiceCollection AddPersistenceServices(this IServiceCollection services, IConfiguration configuration)        {            services.AddDbContext<SmsDbContext>(options => options.UseNpgsql("User ID=makin19;Password=Muhammed1453;Host=81.0.220.136;Port=5432;Database=SMS;"));            services.AddIdentity<User,Role>().AddEntityFrameworkStores<SmsDbContext>();            services.AddScoped<IRefreshTokenRepository, RefreshTokenRepository>();            services.AddScoped<IOperationClaimRepository, OperationClaimRepository>();            services.AddScoped<IUserOperationClaimRepository, UserOperationClaimRepository>();            return services;        }    }}
