@@ -1,7 +1,9 @@
-﻿//using Application.Features.Auths.Commands.Register;
+﻿using Application.Features.Auths.Commands.LoginUser;
+using Application.Features.Auths.Commands.PasswordReset;
+using Application.Features.Auths.Commands.RefreshTokenLogin;
+using Application.Features.Auths.Commands.VerifyResetToken;
 using Application.Features.Auths.Dtos;
-using CoreFramework.Security.Dtos;
-using CoreFramework.Security.Entities;
+using MediatR;
 using Microsoft.AspNetCore.Mvc;
 
 namespace WebAPI.Controllers;
@@ -10,25 +12,37 @@ namespace WebAPI.Controllers;
 [ApiController]
 public class AuthController : BaseController
 {
-    [HttpPost("Register")]
-    public async Task<IActionResult> Register([FromBody] UserForRegisterDto userForRegisterDto)
+    private readonly IMediator _mediator;
+    public AuthController(IMediator mediator)
     {
-        //RegisterCommand registerCommand = new()
-        //{
-        //    UserForRegisterDto = userForRegisterDto,
-        //    IpAddress = GetIpAddress()
-        //};
-
-        //RegisteredDto result = await Mediator.Send(registerCommand);
-        //SetRefreshTokenToCookie(result.RefreshToken);
-        //return Created("", result.AccessToken);
-        return null;
+        _mediator = mediator;
+    }
+    [HttpPost("[action]")]
+    public async Task<IActionResult> Login(LoginUserCommand loginUserCommandRequest)
+    {
+        var response = await _mediator.Send(loginUserCommandRequest);
+        return Ok(response);
     }
 
-    private void SetRefreshTokenToCookie(RefreshToken refreshToken)
+    [HttpPost("[action]")]
+    public async Task<IActionResult> RefreshTokenLogin([FromBody] RefreshTokenLoginCommand refreshTokenLoginCommandRequest)
     {
-        CookieOptions cookieOptions = new() { HttpOnly = true, Expires = DateTime.Now.AddDays(7) };
-        Response.Cookies.Append("refreshToken", refreshToken.Token, cookieOptions);
+        var response = await _mediator.Send(refreshTokenLoginCommandRequest);
+        return Ok(response);
+    }
+
+    [HttpPost("password-reset")]
+    public async Task<IActionResult> PasswordReset([FromBody] PasswordResetCommand passwordResetCommandRequest)
+    {
+        PasswordResetDto response = await _mediator.Send(passwordResetCommandRequest);
+        return Ok(response);
+    }
+
+    [HttpPost("verify-reset-token")]
+    public async Task<IActionResult> VerifyResetToken([FromBody] VerifyResetTokenCommand verifyResetTokenCommand)
+    {
+        VerifyResetTokenDto response = await _mediator.Send(verifyResetTokenCommand);
+        return Ok(response);
     }
 
 }
